style(ALL) : Reformat Code

+++ src/main/java/christmas/domain/Order.java
@@ -1,6 +1,5 @@
 package christmas.domain;
 
-import christmas.constants.StringConstants;
 import christmas.utils.validators.OrderValidator;
 import java.util.Collections;
 import java.util.List;
@@ -8,26 +7,26 @@ import java.util.List;
 public class Order {
     private final List<OrderItem> orderItems;
 
-    private Order(final List<OrderItem> orderItem){
+    private Order(final List<OrderItem> orderItem) {
         this.orderItems = orderItem;
     }
 
-    public static Order from(final List<OrderItem> orderItems, OrderValidator orderValidator){
+    public static Order from(final List<OrderItem> orderItems, OrderValidator orderValidator) {
         orderValidator.validate(orderItems);
         return new Order(orderItems);
     }
 
-    public int requestTotalPrice(){
+    public int requestTotalPrice() {
         return orderItems.stream().mapToInt(item -> item.requestPrice()).sum();
     }
 
 
     public int requestMainCount() {
-        return (int) orderItems.stream().mapToInt(item -> item.requestMainCount()).sum();
+        return orderItems.stream().mapToInt(item -> item.requestMainCount()).sum();
     }
 
     public int requestDessertCount() {
-        return (int) orderItems.stream().mapToInt(item -> item.requestDessertCount()).sum();
+        return orderItems.stream().mapToInt(item -> item.requestDessertCount()).sum();
     }
 
     public List<OrderItem> getOrderItems() {
